split row-level parsing structures out of dwelling
