+++ python/src/energuide/dwelling.py
@@ -1,128 +1,16 @@
 import datetime
-import enum
 import typing
-import cerberus
-from dateutil import parser
+from energuide import interpreter
 
 
 class NoInputDataException(Exception):
     pass
 
 
-class InvalidInputDataException(Exception):
-    pass
-
-
-EvaluationData = typing.Dict[str, typing.Any]
-
-
-@enum.unique
-class EvaluationType(enum.Enum):
-    PRE_RETROFIT = 'D'
-    POST_RETROFIT = 'E'
-
-    @classmethod
-    def from_code(cls, code: str) -> 'EvaluationType':
-        if code == cls.PRE_RETROFIT.value:
-            return EvaluationType.PRE_RETROFIT
-        elif code == cls.POST_RETROFIT.value:
-            return EvaluationType.POST_RETROFIT
-        else:
-            raise InvalidInputDataException()
-
-
-@enum.unique
-class Region(enum.Enum):
-    BRITISH_COLUMBIA = 'BC'
-    ALBERTA = 'AB'
-    SASKATCHEWAN = 'SK'
-    MANITOBA = 'MB'
-    ONTARIO = 'ON'
-    QUEBEC = 'QC'
-    NEW_BRUNSWICK = 'NB'
-    PRINCE_EDWARD_ISLAND = 'PE'
-    NOVA_SCOTIA = 'NS'
-    NEWFOUNDLAND_AND_LABRADOR = 'NL'
-    YUKON = 'YT'
-    NORTHWEST_TERRITORIES = 'NT'
-    NUNAVUT = 'NU'
-    UNKNOWN = '??'
-
-    @classmethod
-    def _from_name(cls, name: str) -> typing.Optional['Region']:
-        snake_name = name.upper().replace(' ', '_')
-        return Region[snake_name] if snake_name in Region.__members__ else None
-
-    @classmethod
-    def _from_code(cls, code: str) -> typing.Optional['Region']:
-        code = code.upper()
-        for region in Region:
-            if code == region.value:
-                return region
-        return None
-
-    @classmethod
-    def from_data(cls, data: str) -> 'Region':
-        output = cls._from_name(data)
-        if not output:
-            output = cls._from_code(data)
-        if not output:
-            output = Region.UNKNOWN
-        return output
-
-
-class _ParsedDwellingDataRow(typing.NamedTuple):
-    eval_id: int
-    eval_type: EvaluationType
-    entry_date: datetime.date
-    creation_date: datetime.datetime
-    modification_date: datetime.datetime
-    year_built: int
-    city: str
-    region: Region
-    postal_code: str
-    forward_sortation_area: str
-
-
-class ParsedDwellingDataRow(_ParsedDwellingDataRow):
-
-    _SCHEMA = {
-        'EVAL_ID': {'type': 'integer', 'required': True},
-        'EVAL_TYPE': {'type': 'string', 'required': True, 'allowed': [eval_type.value for eval_type in EvaluationType]},
-        'ENTRYDATE': {'type': 'string', 'required': True},
-        'CREATIONDATE': {'type': 'string', 'required': True},
-        'MODIFICATIONDATE': {'type': 'string', 'required': True},
-        'YEARBUILT': {'type': 'integer', 'required': True},
-        'CLIENTCITY': {'type': 'string', 'required': True},
-        'CLIENTPCODE': {'type': 'string', 'required': True, 'regex': '[A-Z][0-9][A-Z] [0-9][A-Z][0-9]'},
-        'HOUSEREGION': {'type': 'string', 'required': True},
-    }
-
-    @classmethod
-    def from_row(cls, row: EvaluationData) -> 'ParsedDwellingDataRow':
-        validator = cerberus.Validator(cls._SCHEMA, allow_unknown=True)
-        if not validator.validate(row):
-            error_keys = ', '.join(validator.errors.keys())
-            raise InvalidInputDataException(f'Validator failed on keys: {error_keys}')
-
-        return ParsedDwellingDataRow(
-            eval_id=row['EVAL_ID'],
-            eval_type=EvaluationType.from_code(row['EVAL_TYPE']),
-            entry_date=parser.parse(row['ENTRYDATE']).date(),
-            creation_date=parser.parse(row['CREATIONDATE']),
-            modification_date=parser.parse(row['MODIFICATIONDATE']),
-            year_built=row['YEARBUILT'],
-            city=row['CLIENTCITY'],
-            region=Region.from_data(row['HOUSEREGION']),
-            postal_code=row['CLIENTPCODE'],
-            forward_sortation_area=row['CLIENTPCODE'][0:3]
-        )
-
-
 class Evaluation:
 
     def __init__(self, *,
-                 evaluation_type: EvaluationType,
+                 evaluation_type: interpreter.EvaluationType,
                  entry_date: datetime.date,
                  creation_date: datetime.datetime,
                  modification_date: datetime.datetime
@@ -133,7 +21,7 @@ class Evaluation:
         self._modification_date = modification_date
 
     @classmethod
-    def from_data(cls, data: ParsedDwellingDataRow) -> 'Evaluation':
+    def from_data(cls, data: interpreter.ParsedDwellingDataRow) -> 'Evaluation':
         return Evaluation(
             evaluation_type=data.eval_type,
             entry_date=data.entry_date,
@@ -142,7 +30,7 @@ class Evaluation:
         )
 
     @property
-    def evaluation_type(self) -> EvaluationType:
+    def evaluation_type(self) -> interpreter.EvaluationType:
         return self._evaluation_type
 
     @property
@@ -159,7 +47,7 @@ class Evaluation:
 
     def to_dict(self) -> typing.Dict[str, typing.Any]:
         return {
-            'evaluationType': self.evaluation_type.name,
+            'evaluationType': self.evaluation_type,
             'entryDate': self.entry_date,
             'creationDate': self.creation_date,
             'modificationDate': self.modification_date,
@@ -172,7 +60,7 @@ class Dwelling:
                  house_id: int,
                  year_built: int,
                  city: str,
-                 region: Region,
+                 region: interpreter.Region,
                  postal_code: str,
                  forward_sortation_area: str,
                  evaluations: typing.List[Evaluation]) -> None:
@@ -185,7 +73,7 @@ class Dwelling:
         self._evaluations = evaluations
 
     @classmethod
-    def from_data(cls, data: typing.List[ParsedDwellingDataRow]) -> 'Dwelling':
+    def from_data(cls, data: typing.List[interpreter.ParsedDwellingDataRow]) -> 'Dwelling':
         if data:
             evaluations = [Evaluation.from_data(row) for row in data]
             return Dwelling(
@@ -213,7 +101,7 @@ class Dwelling:
         return self._city
 
     @property
-    def region(self) -> Region:
+    def region(self) -> interpreter.Region:
         return self._region
 
     @property
